Request OpenGL version in iblbaker
The runtimerender module only includes GLSL shaders for prefiltering
for version 150 and 300 es. In line with this, request 3.2 core or 3.0
(ES) in the code too. This would not be suitable for some virtual
machine environments (e.g. due to having an old Mesa llvmpipe only
providing OpenGL 3.0/3.1), but on real hardware >=3.2 should not be a
problem.

Pick-to: 6.1
Change-Id: I18b6f2234127aa50be1d1b8fd5705d46caead934
Reviewed-by: Andy Nichols <andy.nichols@qt.io>