Use new renderpass descriptor cloning to avoid pointers-in-hashmap-key problems
Using importScene presents challenges with lifetimes of objects that are
associated with backend/frontend nodes, because both nodes can outlive
the render target, and can be used with multiple render targets over
their lifetime.

Try preventing problems by creating our own version of the renderpass
descriptor object. This is possible due to the concept of renderpass
compatibility in graphics APIs where this is relevant, see e.g. section
8.2 of the Vulkan spec. Passing a QRhiRenderPassDescriptor backed by a
different VkRenderpass to Qt Quick is fine, as long as the VkRenderpass
is compatible with the one backing the QRhiRenderPassDescriptor from
Qt Quick 3D.

Pick-to: 6.1
Fixes: QTBUG-91888
Change-Id: I222b2805ae7198654512e1c6a5a467cb498a9238
Reviewed-by: Andy Nichols <andy.nichols@qt.io>