Don't crash when View3D is deleted while objects remain
When the view is deleted, the QPointer to the scene manager turns into a
null pointer. Instead of connecting to destroyed() and maintaining that
connection, just assume that the scene manager has been deleted if we
have a null pointer and only one reference.

Fixes: QTBUG-88320
Change-Id: I09444da4f54feb5b383b62be88a133f48332b04d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 55596a475d7d49ab032acffdbf26935c08c7b022)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>