AssimpImporter: Fix to generate a bone hierarchy
The method of assimp's bone identification is just checking if the node
is used as a bone or not. So it is safe to assume all the nodes which
are children of a bone as bones.

Fixes: QTBUG-91407
Change-Id: I7e5ba988b4f9ad0c08f9f50fafc0f69e48d07c2d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 8b0a636e5b6515a88f5d5edd5bc46d63f783a7cc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>