Fix the per frame allocator
Change the alignment to be the alignment of a void pointer, as that
should be a good general purpose value for most of our use-cases.
Note that the old alignment value just happened to work as the objects
we were allocating were all 8 byte aligned.

Change-Id: I90d94bfde4d72046ffb3bc20f1d86ade72079d30
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit d4781d218c70a7d4740a5528c8812763cba6dbac)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>