Asset: Fix the URL handling in the model
From the asset handling side we don't want to do too much magic,
so keep it simple and set the source URL the same way for both usages
of the intermediate scene. For this to work the QQuick3DModel needs to
pass the source url along unmodified if translateSource() fails to
resolve the URL through the various QML URL resolving functions.

Change-Id: Ib12e94a2c3044db21f4db45839a0f0c0173b572c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>