View3D: separate qt.quick3d.hover cat. log from qt.quick3d.pick
A lot of the time, every mouse move generates pick results, but that's
noisy, and was making qt.quick3d.pick hard to use to troubleshoot the
usual press/move/release sequences.  So we avoid logging hover events
to the pick log, and add a separate qt.quick3d.hover logging category.

Change-Id: If4f3992398ac3ce4b2f2e4df227735ece0808e51
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>