Disable input processing for View3D
Input processing of events was used for forwarding events from a View3D
to QQuickItems embedded into the 3D scene. This feature is not quite
mature enough for release, so we are disabling it until it is complete.

Fixes: QTBUG-92603
Change-Id: I21772ef3cf8fa810d435b484e95eb388c8db8eb9
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 4dd8663c64d3b3bf0b7d293e52aa4350c85639ad)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>