Fix the 3d_input manual test
In this test, the relevant DeliveryAgent is usually the subscene agent,
never the main window's DA.

The right stack of cubes is an oddity: they all share textures that come
from the same 2D subscene, which is declared inside of the middle cube,
but re-textured onto the upper and lower cubes.  In addition, the 2D
subscene appears in the lower-left corner of the window if and only if
this texturing occurs at least twice, because of layer.enabled.  So the
same scene appears in up to 4 places, and shares the same subscene
DeliveryAgent instance in all of those.  But the picking logic does not
get confused by this; and ViewportTransformHelper works fine as long as
it is given to THAT subscene.  In this scenario, if we set it on the
main window's delivery agent instead, it gets bypassed during delivery
of mouse updates to the grabber, because those events are going directly
to the subscene DA.  However there may still be a case when we actually
need to set it on the main DA.

If you interact with that stack of cubes though, and then interact with
the other manifestation of that subscene in the lower left corner,
dragging occurs on the wrong coordinate system.  Clicking buttons
and the text field works OK, but not dragging the RGB circle around.
But it's OK if that's the first thing you do.

Change-Id: Ieaa81c6e982f0340d97ea2c3cc686402c3e2715c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 732e59be1052fc8f09f4df4b4ae86ead4fa875c5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>