Square the camera distance when picking in an Item2D subscene
Apparently intersectionTime is actually distance from the camera in
intersectRayWithItem2D(), but QSSGRenderPickResult stores it squared.
This fixes the z-sorting order in QSSGRenderer::processPickResultList()
etc. so that the PickResultList in QQuick3DViewport::internalPick()
will be correctly sorted by z order, and we deliver events top-down.

When a scene contains a combination of Model objects and Item2D 2D
scenes, we need to sort the Z order correctly to avoid the Item2D acting
as if it's on top when it visually is not.

Pick-to: 6.1
Fixes: QTBUG-93095
Change-Id: I80b68837d642bd5e483f6888d59d5a3530759bcf
Reviewed-by: Andy Nichols <andy.nichols@qt.io>