Add position cache modifiers to attractor API
useCachedPositions can be set to false so that attractor shape changes
are accounted for. positionsAmount can be used to change the cache size.

Shape positions are changed to be deterministic based on the particle
index, but the particle index can now be anything from 0 to INT_MAX. The
index is stored on the system, so that different emitters generate
different positions.

Task-number: QTBUG-92391
Change-Id: I378d61b30cc60e1e31ddc1458280bc095adf3fd5
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>