Don't hardcode instance table stride
Remove all knowledge of the instance table stride from the backend, and
let the frontend specify it based on the size of the InstanceTableEntry struct.

It is still not possible to use tables with a different stride, since
the instance table layout is still hardcoded in the shaders. This is
just a cleanup.

Task-number: QTBUG-91456
Change-Id: I956a261c4df88a42146ccbe0a5f29c53ac89b7cc
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>