Don't allow dragging mapped 2D items outside the bounds of their model
In examples/quick3d/dynamictexture, it looked like the doors became a
seamless "screen" when both were closed.  Onto that pair of models is
mapped a ListView.  But neither door projects an infinite plane; so
ViewportTransformHelper::map() cannot know how to map a viewport point
onto a point outside the model.  Therefore we have it return the
last-known-good point if that happens, so when you try to drag the
ListView too far, it just stops.  And because we cannot maintain the
illusion of a seamless screen across two doors, scale them down slightly
to make the "gap" more obvious.

Pick-to: 6.1
Task-number: QTBUG-92944
Change-Id: I8ffa5e8a972580243a2027cf9cc887b0f98c9a51
Reviewed-by: Andy Nichols <andy.nichols@qt.io>