Introduce some civilization to QQuick3DShaderUtilsRenderPass::m_shaders
Make the underlying container for the list property private, and
actually append when asked to append. Provide a separate method to
retrieve a shader by stage.

This is already fixed in a different way in Qt6. Unfortunately the code
there has diverged quite a bit, which makes backporting the change
impossible.

Fixes: QTBUG-93410
Change-Id: I69e62d1c28bb9b601078103443e4b24550225f2b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>