CMake: Prepare repo to mark the ShaderTools package as optional
This partially prepares the repo (from the build system perspective)
to allow the ShaderTools package to be optional (the one that brings
in the QtShaderTools module library).

The ShaderToolsTools package which contains the qsb target
and the qt_internal_add_shaders CMake API is still required
and searched for separately, to support cross-compiling
qtquick3d without having a target qtshadertools build.

When qtquick3d is ready to be built without the QtShaderTools library,
the find_package() call should be changed from REQUIRED COMPONENTS
to OPTIONAL_COMPONENTS.

Task-number: QTBUG-92050
Change-Id: I8ed92e98ffb2216962dc54c288310e36fd34bf6a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>