Shadow UV coordinate variables in Fragment Shader
Some techniques like parallax mapping adjust the value of UV coordinates
in the fragment shader to affect how all other texture reads are
sampled. To make this possible the UV coordinates shouldn't directly
read from the UV varying variable, but rather a shadowed variable that
can be adjusted before use. This change adds this functionality as well
as the ability to delay UV transform calculate until the fragment shader
when necessary.

Change-Id: I3fbe14e5d776566401dd6b350359cfd1b454649c
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>