Fix the occlusionMap to be compatible with GLTF2
The occlusionMap only influences the indirect lighting.

[ChangeLog][QQuick3D][QQuick3DMaterial/OcclutionMap] This patch changes
how occlusionMap works in QQuick3DMaterial. Previously occlusionMap
influences on all the diffuse colors and emissive colors. Now it will
not affect punctual lightings.(For now it will affect only IBL.)
The result of a lancelot test, principled/occlusion.qml, will be changed
because it does not have IBL.(The occlusionMap will be ignored.)
There will be 4 rows in the new test. Top 2 rows are without IBL, and
others are with IBL in their materials.

Pick-to: 6.1
Change-Id: I0087e10d711eb14917050d9c1447c769d811fc3e
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>