Add support for setting depthDrawMode on Materials
This makes it possible to define when a material writes to the depth
buffer during rendering. Previously the behavior was that only opaque
items could write to the depth buffer.  There are cases however for
correct rendering it is necessary for individual materials to define if
and when they will write to the depth buffer.

Task-number: QTBUG-92894
Change-Id: I184d1f04c7d976b8ad91920360cda7e760a7e105
Reviewed-by: Christian Strømme <christian.stromme@qt.io>