Allow using a QQuick3DModel as an emitter shape
The emitter shape can be set to a model in the scene which causes the
emitter to emit particles in the shape of that model.

When fill property is false, random position from the surface of the
model is selected for each particle. When fill is true, random position
between a random surface position and the center of the model is chosen.
This fill method works more accurately for convex models than concave
ones.

Task-number: QTBUG-91471
Change-Id: I03d985ad8fed9def580f6b1b963d0212c2de04d4
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>