Remove unused arg in loadCompressedImage
Just some internal pedantic cleanup: the argument is unused because
it cannot be used for anything when working with compressed texture
data. An argument defaulting to true that is then unused (and so
behaving as if it was false) is rather confusing looking so just
drop it.

Change-Id: Ib94760d6a52f5049b99ad378fa545f807ad39450
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 99d1f457abfa37fa300807b03e8440654cdf81ec)
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>