Enable shadowing of transparent models with OpaqueDepthPrePass mode
Previously only materials which were considered opaque could contribute
to the shadow maps.  Now if a transparent material is using the
depthDrawMode of OpaquePrePassDepthDraw, it also will contribute to the
shadow maps.  Only fragments that have an alpha value of 1.0 will
contribute though.

This follows the behavior of other rendering engines.

Task-number: QTBUG-92894
Change-Id: If82659262f51b447793f751334da565f6c690de4
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>