Refactor shared code into addOpaqueDepthPrePassBindings method
For the depth and shadow passes there is a block of common code for
binding the necessary texture samplers for calculating the alpha value.
This code has now be refectored into a new method.

Code has also be added to ensure that that all samplers will be bound to
textures, and when they are not defined in custom materials, dummy
textures will be used like in the main render pass.

A lancelot test has also be added to test the dummy texture mechanism in
all paths (color, depth, shadow).

Change-Id: Ieb22ff7543e351d1294b806d4de8aec6241194d8
Reviewed-by: Christian Strømme <christian.stromme@qt.io>