Asset: Move scene creating code out of assimp plugin
This part of the code is not tied to assimp at all, the only reason
it was in the plugin was for convenience. The code is now moved
into assetutils the same way the qml generator is.

Change-Id: Ibd833ce0bd6b7b24d87d8b23be107eaa648339cf
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 457776dc9faa0b6b67f94c3ac9062492eee03aac)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>