Expose local position and normal to pick result
We have already gone through the effort of calculating the normal and
local position but we didn't pass it through to the PickResult. This
just makes sure that these are passed all the way to the front end
because these values a useful for interactions.

This change is part of the input API review.

Pick-to: 6.2
Change-Id: I80a7b30984498f43fb3083bd3f143270ac6fdbce
Reviewed-by: Christian Strømme <christian.stromme@qt.io>