Add another environment map test for Lancelot
Exercises Texture.sourceItem and also has a somewhat dummy skysphere,
thus demonstrating using Qt Quick generated content both for the
background (skybox) as well as an environment map to get reflections.

The image quality in the output is not that great because the image
is resized to 512x512 instead of 1024x1024 to keep the file size low.

Task-number: QTBUG-94453
Pick-to: 6.2
Change-Id: I685c6066c1f463ddcde2cf501bc93e2c38d3e4dd
Reviewed-by: Andy Nichols <andy.nichols@qt.io>