Revert "Always fetch the assimp submodule when fetching QtQuick3D"
This reverts commit b22ce016473ef2b0ec8ec0705cf3f88b11f5f40d.

Reason for revert: The option does not cause the submodule to be
checked out by default, as the commit message implies. Instead,
it just causes "git fetch" to always run git fetch also in the
submodule. The default behavior - "on-demand" - is to fetch the
submodule only if the fetching of qtquick3d.git retrieves an
update involving an update of the submodule, which should be enough.

Pick-to: 6.2
Change-Id: I468dc0b3f1fd3c5e3b81b21e695e90ba771a3e36
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>