Add graph type for QQuick3DFrustumCamera
This one was missed in 0908daf99cb, so this change  adds a type for the
frustum camera as well. Note that the frustum camera has been piggybacking
on the perspective camera's type, and since we are still using separate
projection flag for the camera backend nodes, there's not been a real need
for using the type.

Pick-to: 6.2
Change-Id: I0fb4bd6713c77b6459b8e943e193057b6345e1ff
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>