Remove runtime flags for setting the camera projection
The camera already has the type set when it's created, and the type
shouldn't changed after the node has been created, as that's
completely unreasonable thing to support here. The graph object type
also implies the projection, so there shouldn't be two ways to set it,
as was the case.

Pick-to: 6.2
Change-Id: Id16be909aad4127da26906e1f2affa9db40c6836
Reviewed-by: Andy Nichols <andy.nichols@qt.io>