CMake: Work around command line length issue on windows hosts
CMake 3.21.0 is now passing all include paths as absolute paths to the
compiler.

This causes command line length issues on windows hosts when
building the assimp importer plugin with Ninja due to the large number
of include paths and defines passed.

Work around the issue by using response files just for this target and
the resource target.

Allow opting out of this behavior by passing a
-DQT_NO_NINJA_FORCE_RESPONSE_FILE=TRUE option when configuring the
module.

Qt's CI uses sccache for compilation caching and that also causes the
command line length issue. Presumably sccache reads the contents of
the response file and passes all the options verbatim to the
compiler. That will obviously hit the limit despite the response
file usage.

Disable the sccache compiler launcher for the plugin importer
target and it resource target.

Pick-to: 6.2
Task-number: QTBUG-95284
Change-Id: I8ec8cc9ede6faabed8933bfef6f68b7991b59810
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Craig Scott <craig.scott@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>