Remove conditions for tools
The host tool magic is in qt_internal_add_tool, that ensures the tool is
not built for the target, but rather does find_package to find the tool
from the host Qt.

Skipping tools like balsam and shadergen, which are then used in our
CMake Macros files, leads to breaking our CMake functions in cross
compiled builds due to CMake not knowing anything about tools/targets
like Qt6::shadergen.

If there are applications under tools that really should not be built
for some platform, then let's reintroduce conditions on top of this
patch. The old condition ported from the .pro file was questionable in
the first place.

Pick-to: 6.2
Fixes: QTBUG-95535
Change-Id: Ic01ba1b9b35d763a7b263e5c6087762068f9f6fc
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>