Fix z-prepass rendering for transparent objects with depth write
The rendering wasn't done correctly when the scene has only transparent
objects, but some of them have depth-writing enabled and scene has
the z-prepass enabled. The zprepass was doing the depth writing,
but the shading pass didn't enable depth testing so it had no effect.
The depth-testing wasn't enabled because the depth-testing gets disabled
when there are only transparent objects if they don't write to depth
buffer, but this wasn't properly checked for all objects.

Fixes: QTBUG-95586
Pick-to: 6.2
Change-Id: Idcfcecc473a622bf106c37466bf7476b3d1fd570
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>