Avoid race condition in picking code
This patch addes a mutex to the RenderBufferManager to prevent a race
condition when picking occurs at the same time that a custom geometry
buffer is being updated. What happened was we were using the buffers
data for picking, but in the mean time that buffer was deleted and
replaced with an updated buffer. This can happen quite easily if a
custom geometry item was updated as a result of a pick, and another pick
happens before the next frame is rendered.

Fixes: QTBUG-94335
Change-Id: I15aa2449a337ba8ad33aed32611dcff6cd01465b
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 99772198f0046cf208d5e4eddc0057c8809b16a6)
Reviewed-by: Andy Nichols <andy.nichols@qt.io>