Change the logic for how we pick which camera to use for rendering
If a camera has been explicitly set by the user and that camera is _not_
active (visible), then we would fallback to picking the first active
camera in the scene, this means we're completely ignoring that the user
actually wanted to disable the camera and not having it magically pick
some other camera in the scene. With this change we'll only go looking
for cameras if a camera has not been explicitly set by the user.

Change-Id: I4fb56b597bc846c0621f336dc69dccec633c1f91
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit bef123136d294d1eeb81adb97f26f5abd1c51fd7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>