Enhance add_lightprobe_images file name handling
[ChangeLog] The qt6_add_lightprobe_images CMake function, which is the
new name of qt6_quick3d_bake_lightprobe_hdri has been changed to match
qt6_add_resources when it comes to the handling of PREFIX, BASE, and
FILES. It no longer strips the directories from the entries in the FILES
list. This means that invocations that previously catered for the special
behavior by listing the necessary extra directories in PREFIX must now be
changed to not do that anymore. In addition, support for the OUTPUTS
keyword has been introduced, in order to have API symmetry with
qt6_add_shaders.

Pick-to: 6.2
Change-Id: I20c7e8834cf03455084be013b7d628860881477d
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>