Simplify QML module definition and remove CMake warning
This removes a warning about using the QT_QML_SOURCE_VERSION source
property, which was recently renamed. It is more appropriate to set
the module's PAST_MAJOR_VERSIONS instead, which means the source
property doesn't have to be explicitly set any more.

Also fix the DEPENDENCIES to use auto for the version as a drive-by.

Pick-to: 6.2
Change-Id: Icf320faa7768e8f77b0d4af765c44736e0be844d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>