Assimp: Add handling for FBX Morph Animations quirks
The target name for animations (both keyframe and morph) animations can
contain wildcard characters, meaning that it is possible for a
definition to apply to multiple nodes. Currently when the name contains
such a wildcard character we don't apply it to anything because we just
compare the string names directly. So this patch tries to search for at
least one item that meets the search criteria.

The assimp FBX importer will always hit this path because no matter what
the target name is, it always applies "*" + the target geometry index
value. So it would appear as "name*0" instead of just "name".  This
patch also makes sure to strip off this final index value as it is not
currently needed.

So this patch should fix animation imports of morph targets for FBX
files, but does leave some open questions about supporting intentional
wildcard names. But for now we just always take the first match, which
is better than always failing.

Pick-to: 6.2
Change-Id: I6c0f8a69fc4c10ca8d2760a5c4922f3d27a48fd3
Reviewed-by: Christian Strømme <christian.stromme@qt.io>