Make sure to install designer files with non-prefixed builds
It is not enough to just call qt_copy_or_install with a DESTINATION
value prefixed with only ${INSTALL_QMLDIR}/ because for non-prefixed
builds this will install relative to the directory this cmake file is
in. Instead we have to combine ${QT_INSTALL_DIR} with the target path to
make sure the final path is always correct. This pattern is used all
over, so we muse do it as well.

Change-Id: Ia5877e321b3d8ec0fc6325ca4dfc54c4bbd878cf
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit a7c2ff838b4f246fd8c1cf43b00286458d76e40c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>