Make the "sharedSource" input autotests actually share the source
It was a misnomer.  The idea is to have one 2D subscene used as
sourceItem in two different Textures mapped onto two different Models.
And now the dualTouchTap2D test needs to press two different buttons at
the same time with the two touchpoints, rather than two instances of
the same button.  But there's only one subscene DeliveryAgent (plus the
window's DA), and only one instance of ViewportTransformHelper is
created; so we cannot yet support the use case of pressing anything
in the same subscene mapped onto two Model objects at the same time,
because we don't have the ability to remember how to redo the two
separate viewport->3d->subscene mappings after the touch press.
ViewportTransformHelper only remembers one QSSGRenderNode.
But pressing two different buttons in the same subscene is ok, even in a
real shared-source case, as long as the user only tries to press the two
points on the same 3D model at the same time, not across two models.

This isn't reproducing QTBUG-92841 yet, but it's necessary to have
a shared-source test case first.

Pick-to: 6.2
Task-number: QTBUG-92841
Task-number: QTBUG-96324
Change-Id: I0b25db3b65f225ac2f1fae3f7b56b674094e504e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>