Assimp: Only the first animation timeline should be enabled
When importing assets with multiple animation timelines, it doesn't make
sense for all animations to play at the same time.  So instead we can
keep the default behavior of automatically playing the first timeline,
but disable the rest. This works well with how Design Studio works with
animations because then when associating animation timelines with
differnt states they get automatically enabled/disabled.

Pick-to: 6.2
Change-Id: I73e38dbd3e9f17119783326b2458cd0a6830ecdd
Reviewed-by: Christian Strømme <christian.stromme@qt.io>