Pull keyboard focus back to WasdController when clicking or dragging
If the user clicks on some 2D item that takes focus, it tended to get
stuck there, because there wasn't a way to click "outside" that item
to give it back to the WasdController so that keyboard navigation
would work again. Now when you either click or drag in such a way
that the events fall through to the WasdController (which is usually
used underneath a View3D), it calls forceActiveFocus().

We could use Shortcut to avoid needing focus at all; but we can't just
take over the WASD keys completely, because then you couldn't type those
characters in any text fields. So it's hard to imagine any other
alternative than click-to-focus.

Fixes: QTBUG-79026
Change-Id: I5c053bf3cef1b77b38629a31c6f791a6aeb1d304
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 51c266558cffef60057fd0003b35edc56337e65b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>