QQuick3DViewport::internalPick: ignore empty Item2D subscenes
If a 2D item is declared inside a Model or Node but also has a parent:
declaration to reparent itself somewhere else, an empty 2D subscene gets
created before the reparenting is done, and continues to exist, but has
no children after the reparenting. There's no need to deliver input
events to an empty subscene.

Pick-to: 6.2
Task-number: QTBUG-96457
Task-number: QTBUG-96514
Change-Id: I3a5b91ecf89bb4583b039af2b7bf534fbf8e5878
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>