QQuick3DViewport::internalPick: ignore hits outside Item2D.childrenRect
An Item2D is an infinite plane for rendering purposes: e.g. if it emits
2D particles, we don't want those to get cut off at some invisible edge.
But for event delivery purposes, it should not block events from
"falling through" to other content behind this infinite plane; so we now
ignore pick results whose UV coordinates indicate that they fall outside
the childrenRect of the subscene root.  The subscene root itself has
zero size, so it's not enough to check its clipRect() or boundingRect().

Pick-to: 6.2
Task-number: QTBUG-96457
Change-Id: Ie384d6d613107302cc5f3548c297b9472e5a8403
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>