Retire invalidateCachedReferences
Rest in peace. Now that it is possible to extract an opaque blob from
both QRhiShaderResourceBindings and QRhiRenderPassDescriptor describing
the binding list layout and the render target setup (color attachment
count, their formats, depth/stencil presence, etc., whatever is relevant
with the underlying QRhi backend and graphics API), we can store and
compare these instead of the objects themselves and calling member
functions such as isLayoutCompatible() or isCompatible().

This way we avoid storing object references in data structures used as
keys to maps and hash tables, and the headache caused by having to check
and remove entries where the key refers to an about-to-be-destroyed
objects disappears.

Additionally, in case of a graphics pipeline keeping pipelines around is
beneficial even: there is a high chance that the same pipeline can be
reused later, even if the render target('s rp descriptor) with which the
pipeline was originally associated with is destroyed; an application
typically works with a few type of render targets, with most falling in
the same category as many others, the most common example being 1 color
attachment, where attachment 0 is an RGBA8 texture, 1 depth-stencil
attachment with a format like D24S8, etc. A pipeline built with such
a target can be reused with any other targets of the same setup, which
is what previously we achieved by calling isCompatible() on the
QRhiRenderPassDescriptor, it's just that this system can be improved
by eliminating the need to call a member function, so that we do not
need to keep the QRhiRenderPassDescriptor around once the pipeline is
successfully built. This also eliminates the need to remove an entry
from the cache just because the key is about to have a dangling pointer
in it.

The QRhi APIs conveniently return implicitly shared containers (which
are pre-calculated upon baking the object, so the getters here are
cheap, and it's also cheap to copy these and store them in the structs
that serve as keys for the hash tables.

Change-Id: Ia8248be2ddc24ee4199f8cfe408d1a930f970ae9
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 23d71a1de6a6f1fa740ad8984737685965ec342e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>