Insert line directive for custom and effect shaders
Inserting the line directive to user supplied shaders to make the
error reporting more useful. If QT_RHI_SHADER_DEBUG is set to a greater
then 0, then the line numbers will not be adjusted, as the fully expended
shader(s) will be printed and not just the users shader code.

Fixes: QTBUG-96736
Pick-to: 6.2
Change-Id: I3cc54765a507c985df54cf0de7d8e18c2d5bf0c2
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>