Improve feature list handling for shader generation
Don't sort, search, hash, re-hash and cache the feature set headlessly.
Not only did we do pointless gymnastics to create a key based on
the features, but we stored them as separate entities taking up extra
space for no obvious reasons, which also made copying the state from
one place to another non-trivial. The features are now stored in
a single 32 bit integer that uniquely identifies the feature set and
can be used to get the state, name (define) and enumerator for all
its values.

This change also makes the shader features in the shaders consistent,
meaning: A feature is defined to be 0 by default at the top of the
shader, which means the default behavior should work. Testing if a
feature is enabled '1' or disabled '0' should be done for the preceding
code. Furthermore, the preprocessor directives defining all the feature
used by the shader pipeline and their state will be written out in the
header of the generated code, and not just for the ones that will be
used (enabled).

Change-Id: I51fe16c2d08ed2847c4e45071ae17049434434cb
Reviewed-by: Andy Nichols <andy.nichols@qt.io>