Add reflection probe
Adds support for a reflection probe object. It renders the scene to a
cube map that is then used to render reflections to objects in a
similar fashion as the light probe. Only real-time rendering is
supported in this commit.

Co-authored-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Task-number: QTBUG-97960
Task-number: QTBUG-97954
Task-number: QTBUG-97955
Task-number: QTBUG-97956
Task-number: QTBUG-97957
Task-number: QTBUG-97958
Task-number: QTBUG-97959
Change-Id: Ia6cc8029cb783328e91cbb80909a4e61c64a878e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>