Fix shadow map bounds
The major change in this commit is that we use the bounds of objects
casting shadows to calculate a bounding box and then expand this in the
forward direction to also cover all objects receiving shadows. We also
compute the camera transformation matrix by hand since the camera lookAt
method is broken. There is also some minor cleanups and refactoring.

Change-Id: I25bbe5487b1f37a8d23f9fa22d2c55365cabcba6
Reviewed-by: Andy Nichols <andy.nichols@qt.io>