Remove Render Resource Manager
This code didn't really do anything other than wrap the creation of a
QRhiTexture and QRhiRenderBuffer in practice. The state goal of the
class was to create a pool of "free" textures which could be reused
between frames, but this was not used because the textures were never
released to this cache, and was only used by the shadow code which does
not deallocate between frames anyway.

This change also removes the m_windowDimensions member from ContexCore
as well all the layer prep code that passed it around because this size
was only used to reset buffers cached by the resource manager being
removed.

Change-Id: I2191409b56b213ac170fcd96efa529e8d2024371
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>