Remove redundant lookup in QSSGBufferManager::loadRenderImage
We would not hit this conditional if { image->m_imagePath, inMipMode }
was found in imageMap, so there is no point in calling find again. At
some point this must have been different because the result was being
used to determine if transparency needed to be determined or not.  In
practice it would always check, so we keep doing that.

Change-Id: Ib1bdccd4f3ddbd83daa8091542aebaa02770ba63
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>