RenderBufferManager: Track QSG texture usage per frame
Previously we did not keep track of QSGTexture usages, so when a
QSGTexture was no longer in use, we still kept a handle to the released
texture. Since we didn't create or own the texture, this didn't show up
as a significant leak, but we do track the texture handle as well as a
bit of meta-data so it would add up for certain use-cases eventually.

This change is also an experiment with per frame tracking of resources.

Change-Id: I46f764b653192ef9ecea57654e7a4b93fa9b3cd1
Reviewed-by: Christian Strømme <christian.stromme@qt.io>