RenderBufferManager: Switch all texture buffers to per frame tracking
This change also make sure that all texture buffer references are
queried from the buffer manager, and not stored in the scene graph since
there can be multiple renderers sharing the same scene graph, thus
multiple render buffer managers.

QSSGRenderTextureData needed to be reworked since it was storing texture
data that might be invalid between two buffer managers. The dirty
mechanism has been changed from a bool to an incrementing value for each
generation.  If the generationId doesn't match with the buffer in the
buffer manager, a new one will be generated.  But this does mean that we
can not release the data buffer when the texture is uploaded.  This may
need a more clever solution later to prevent all these buffer copies for
textures.

Change-Id: Ibff37bd722b86faf31a0bf8a9c8ac29ac0f1b29c
Reviewed-by: Christian Strømme <christian.stromme@qt.io>