Improve handling of materials that depend on the screen texture
This change refactors how the screen texture is generated so that:
a) The backgroundMode of the SceneEnvironment is used to "clear" the
screen texture instead of always clearing with transparent (vec4(0.0)).
b) Using SCREEN_TEXTURE does not implicitly set the hasTransparency
flag, because not all techniques using using the screen texture require
blending.
c) When sorting renderables into opaque and transparent items, there is
now a 3rd list just for items requiring the screen texture. This makes
it easy to skip rendering them when creating the screen texture (instead
of just assuming they are always transparent), but also enables them to
be rendered at the correct time and order: after the opaque items,
sorted from back to front.

Currently this change only affects CustomMaterials, but lays the
groundwork needed to enable the use of the screen texture pass with
DefaultMaterial based materials to enable non "alpha as coverage"
transparency techniques.

[ChangeLog][QQuick3D][QQuick3DCustomMaterial/SCREEN_TEXTURE] This patch
changes the content of the screen texture. The intention of the screen
texture is to enable a material to bind the contents of the framebuffer
currently being rendered to, and that should include the background as
well. Previously the background of the screen texture was always
transparent (alpha = 0.0), but now this will depend on the
backgroundMode set in SceneEnvironment.

Change-Id: I597b0363c8be38ed2d7a915ce47f704f080342a0
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>